requirements.txt
kubernetes>=28.1.0
click>=8.1
jinja2>=3.1
pyyaml>=6.0
